+++ CHANGELOG.md
@@ -6,9 +6,14 @@
 
 - `ModelArrows`
 - `ModelText`
+- `Solid.Intersects(Plane p, out List<Polygon> result)`
+- `Vector3.IsUnitized()`
+- `Transform.Inverse()`
+
 ### Changed
 
 ### Fixed
+
 - Deduplicate catalog names during code generation.
 
 ## 0.9.2
@@ -43,7 +48,6 @@
   - `Line.ExtendTo(Polygon)`
   - `Line.ExtendTo(Profile)`
 - Support for DXF from many basic elements.
-
 
 ### Changed